+++ Source/BLE.Client/BLE.Client/ViewModels/DeviceListViewModel.cs
@@ -228,8 +228,8 @@ namespace BLE.Client.ViewModels
             _cancellationTokenSource = new CancellationTokenSource();
             RaisePropertyChanged(() => StopScanCommand);
 
-            Adapter.StartScanningForDevicesAsync(_cancellationTokenSource.Token);
             RaisePropertyChanged(() => IsRefreshing);
+            await Adapter.StartScanningForDevicesAsync(_cancellationTokenSource.Token);
         }
 
         private void CleanupCancellationToken()